+++ enb/config/aoptions.py
@@ -259,6 +259,13 @@ class DirOptions:
         """
         _singleton_cli.WritableOrCreableDirAction.assert_valid_value(value)
 
+    @OptionsBase.property(action=_singleton_cli.ReadableDirAction, default=calling_script_dir)
+    def project_root(self, value):
+        """If set, data paths relative to the invoked script's paths are stored instead of their
+        absolute path counterparts.
+        """
+        _singleton_cli.ReadableDirAction.assert_valid_value(value)
+
 
 @_singleton_cli.property_class(OptionsBase)
 class RayOptions:
@@ -452,10 +459,13 @@ def propagates_options(f):
     """Decorator for local (as opposed to ray.remote) functions so that they
     propagate options properly to child workers.
     The decorated function must accept an "options" argument.
+    Furthermore, the current working dir is set to the project root so that
+    any relative paths stored are correctly handled.
     """
 
     @functools.wraps(f)
     def wrapper(*args, **kwargs):
+        os.chdir(kwargs["options"].project_root)
         set_options(kwargs["options"])
         return f(*args, **kwargs)